02_RestAPI_server/requirements.txt
boto3==1.18.21
botocore==1.21.41
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.8
Django==2.2.5
djangorestframework==3.12.4
filelock==3.3.0
h5py==2.10.0
huggingface-hub==0.0.19
idna==3.2
importlib-metadata==4.8.1
jmespath==0.10.0
joblib==1.0.1
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
mysqlclient==2.0.3
nltk==3.6.5
numpy==1.21.2
olefile==0.46
packaging==21.0
Pillow==8.3.1
pip==21.0.1
pycparser==2.20
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
s3transfer==0.5.0
sacremoses==0.0.43
scikit-learn==1.0
scipy==1.7.1
sentence-transformers==2.1.0
sentencepiece==0.1.96
setuptools==58.0.4
six==1.16.0
sqlparse==0.4.1
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.9.1+cpu
torchaudio==0.9.1
torchvision==0.10.1+cpu
tqdm==4.62.2
transformers==4.11.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wheel==0.37.0
zipp==3.6.0
